Organize files for first 10 chapters.
The xml files are currently poorly named and do not make it easy to find the
content that you need to edit.  This is a pass and making the file names
match the chapter names.  This is the first of several CRs that, together,
will cover all of the chapters.

Change-Id: I511720eff6a2db0e2a009379222d42686aee65f9