Shorten case studies section titles
Remove the part after the colon for most case study section titles. This
information is now redundant with the chapter title that they fall under.
Keep the colon / subtitle for the one case study that exists as a stand
alone chapter as that still provides a useful purpose. Once that one is
split into the preceeding chapters, it can also follow this new convention.

Change-Id: I36f905abeb4c4e4278fdc25d1850c38629a20619