Fixing wording in the content for Key management
The sentence that describes the provider liablity aspect
of it was not worded properly and this bug corrects
the sentence wording

There was a typo in the bug id and an additional number
got into the Bug number. Fixed it.

Changed the sentence wording as recommended.

Modified the sentence as per recommended grammar rules

Did the git rebase and committing the change

Closes-Bug: #1342296

Change-Id: Iafd5e324d3bd281285cf28fc07059f38aeb33910