Making section id names readable in security guide
In the security guide, the section id names are the chapter name
with a random id number on the end. This is harder to work with and not
really necessary, for example,
case-studies-system-documentation-idp44480. Replacing section id names with
"case-studies-system-documentation-alice-private-cloud".

Changing the ids at the section definitions for case studies throughout.
No references are made to these section ids within security guide and
hence no changes needed there.

Partial-Bug: #1340388

Change-Id: Ibaccd164332239dc9d701a74f5b84c2fd357feb6