Making section id names readable in security guide
In the security guide, the section id names are the chapter name
with a random id number on the end. This is harder to work with and not
really necessary. Replacing random id number with actual section names.

Changing the ids at the section definitions for all the sections throughout.
References to these ids are also updated.

Change-Id: Id4aa5df8c7c025ae7a812f1b031041b14002c030
Closes-bug: #1340388