Updating Case Studies - Alice's Monitoring & Logging
Updating Alice's Monitoring & Logging:
+ SOC analyst to monitor events
+ SEIM already exists
+ Ensures rest of SOC doesn't have access
+ Service groups get own views
+ Clarity around SOC analyst and trimming whitespace

One gap is specific events. As soon as this is merged, I will open a
bug against specific log types being tracked and will highlight what
events Alice is looking for each component, however I expect the
research on that to take some time and this should be put up in the
mean time.

Change-Id: I42e5b6e214025cda309a525bff5f8a805d33b664
Partial-Bug: #1349540