Run checkniceness/doc8 against security-guide-rst
Previously we weren't running doc8 against
security-guide-rst which means that it wasn't being checked.

This fix also required changing 4 of the rst files already
merged as they had incidents of line-length exceeding the
maximum, which is now caught by doc8.

Once the conversion is complete this entry must be reversed
as the project will be renamed.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>
Change-Id: Iebceeeac67406519a49a8f892f636919efc114c3