Updating code block
Policy file and token sample were not merged as expected.
Replacing links with the partial content as a sample.
We can replace the content back in future with the links
if we find a solution for it.
Otherwise, keeping content as is is reasonable in this
context.

Change-Id: Ic25709bf264cfa298fe1fc612629d40b009d85a9
Partial-Implements: bp sec-guide-rst