There's an grammar error in section "Compartmentalize" of Compliance overview in Security Guide.
In section "Compartmentalize" of this URL, there's the following
description:
"Systems should be segregated in a such way that if one machine,...".
"in a such way" is not right, it should be "in such a way".
The suggested description is as following:
"Systems should be segregated in such a way that if one machine,..".

Closes-bug : #1491041

Change-Id: I0e531ec59efcd5223e7eb0758f43d512b798f6a1