Update Links in Management interfaces in Security Guide
In the Management Interfaces section of the Security Guide,
the two links under Management Utilities that redirect to the
User Guide do not correctly go to the sections that are
referenced in the Security Guide.

This fix is to update the links to point to the Command Line Client
Overview and the Download and Source the OpenStack RC file section.

Change-Id: I30a9cbafe5c6c5890401156995dc84c4a1c2f946
Closes-Bug: #1490047