Add marker files
Add marker files in the root of each document.

This is needed for publishing in the future:
This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior builds.

This allows those superior builds to be rsynced without removing the
other builds that were published under them.

Change-Id: Id0f63749e767c9e9db9619d2cb9320b1a20f5995