OSSN-0087: Ceph user credential leakage via Manila
It is possible for regular users of manila
to obtain Ceph client keys that they shouldn't
have access to. This vulnerability occurs because
of a flaw in a ceph interface that manila
interacts with. The flaw has now been patched in
several stable releases of ceph. This security note
is to socialize the fix among OpenStack Manila
deployers so they can understand the vulnerability
and implement the fix in their environments

Closes-bug: #1904015
Change-Id: I911212ea1147b5c3d7ab80835a165cf47c343f6e