Correct the scope of OSSN-0090
Corrected the scope of "Discussion" section from limiting it to
end-users like outlined in the bug comment #43 [0].

Removed the "hence" from line 86 as that would be suggesting
Glance doing the checksumming normally, which is false impression..
The data is not verified because of not going through Glance
but because the consumer decides to not verify it. Subtle but
important difference.

[0] https://bugs.launchpad.net/glance/+bug/1990157/comments/43

Change-Id: Ib42b486f854e39cdae8762f596266d6c24e8b3fb